requirements.txt
# Ethereum and Web3 libraries
web3==6.11.3
eth-account==0.9.0
eth-utils==2.2.2

# Data analysis and manipulation
pandas==2.2.0
numpy==1.26.3

# Database
sqlalchemy==2.0.25

# Graph analysis and visualization
networkx==3.2.1
matplotlib==3.8.2
plotly==5.17.0
graphviz==0.20.1

# HTTP requests and async
requests==2.31.0
aiohttp==3.9.1

# Logging and configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Development tools
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Progress bars and CLI
tqdm==4.66.1
click==8.1.7 